src: app_main.c: Stop catching throw in app_main()
The app used to throw errors in "nominal" situation, including user deny
of a request or even user approval. This was at that time seen as an
efficient way to answer the APDU while not having to forward the error
code.

This was a tricky mecanism, relying on an even trickier TRY/CATCH
mecanism which brings many potentials issues.

Now, errors can be throwned only from the OS/SDK and in non nominal
situation. So it is safe to stop catching theses errors here, which
allowed to send a response to APDU, and let them be catch in the main()
function which is in SDK standard_app library.